Implement the `once` method
Register a callback that is immediately removed after it is invoked the
first time.

The passed callback is wrapped in a function that is registered for the
event. When that wrapper function is called, it removes itself from the
registry, and calls the original callback.

To support removal with the `off` method, the wrapper function is
attached to the original callback. If it is ever passed to `off`, its
wrapper function can be read.