Remove reliance on `Object.keys` in `mixin`
Need the emitter to work in environments where ES5 Shim cannot be
included due to file size considerations. This is the first change on
the path to removing the ES5 Shim dependency.

Use `Object.prototype.hasOwnProperty` in case the object getting
`Emitter` mixed in was created with `Object.create(null)` and has a
`null` prototype chain (i.e., no `hasOwnProperty` method).